refactor: consolidate VFX batch settings contract
- Move VFXBatchSettings ownership to VFXBatch with the other batch contracts
- Replace duplicated StoredBatchSettings fields with a type alias
- Keep a BatchedRenderer type re-export for existing imports
- Update batch, particle system, and test imports
